pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ogc_landing"
version = "0.1.0"
description = "AWS Lambda functions for OGC Landing"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "Michael Leedahl", email = "michaelleedahl@example.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "boto3>=1.20.0",
    "aws-cdk-lib>=2.0.0",
    "constructs>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["ogc_landing"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
